requirements.txt
Django==1.6.1
argparse==1.2.1
django-nose==1.2
factory-boy==2.3.0
nose==1.3.0
six==1.5.2
spec==0.11.1
wsgiref==0.1.2
